fix(test): consolidate SES MIME boundary tests into EmailService.test.ts
